+++ src/audio_core/hle/hle.h
@@ -77,8 +77,8 @@ private:
     void OutputCallback(s16* buffer, std::size_t num_frames);
 
     std::unique_ptr<Sink> sink;
-    std::atomic<bool> perform_time_stretching = false;
-    std::atomic<bool> flushing_time_stretcher = false;
+    std::atomic_bool perform_time_stretching = false;
+    std::atomic_bool flushing_time_stretcher = false;
     Common::RingBuffer<s16, 0x2000, 2> fifo;
     std::array<s16, 2> last_frame{};
     TimeStretcher time_stretcher;
