+++ src/citra/hotkeys.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>hotkeys</class>
- <widget class="QWidget" name="hotkeys">
+ <class>Hotkeys</class>
+ <widget class="QWidget" name="Hotkeys">
   <property name="geometry">
    <rect>
     <x>0</x>
